+++ src/dune/action_plugin.ml
@@ -1,3 +1,4 @@
 let syntax =
   Dune_lang.Syntax.create ~name:"action-plugin" ~desc:"action plugin extension"
+    ~experimental:true
     [ ((0, 1), `Since (2, 0)) ]