+++ test/blackbox-tests/test-cases/default-implementation/package-mismatch1/package-mismatch1.t
@@ -1,0 +1,29 @@
+A default implementation of a library must belong to the same package
+
+  $ echo "(lang dune 2.6)" > dune-project
+  $ touch dummyfoo1.opam
+  $ touch dummyfoo2.opam
+  $ mkdir vlib impl
+  $ touch vlib/vlib.mli impl/vlib.ml
+  $ cat >vlib/dune <<EOF
+  > (library
+  >  (name vlib)
+  >  (modules vlib)
+  >  (public_name dummyfoo1.bar)
+  >  (virtual_modules vlib)
+  >  (default_implementation def_i))
+  > EOF
+  $ cat >impl/dune <<EOF
+  > (library
+  >  (name def_i)
+  >  (public_name dummyfoo2.bar)
+  >  (implements dummyfoo1.bar))
+  > EOF
+  $ dune build @install
+  File "vlib/dune", line 6, characters 25-30:
+  6 |  (default_implementation def_i))
+                               ^^^^^
+  Error: default implementation belongs to package dummyfoo2 while virtual
+  libarary belongs to package dummyfoo1. This is impossible
+  
+  [1]