+++ test/blackbox-tests/test-cases/github3412/run.t
@@ -1,8 +1,0 @@
-  $ dune build
-  File "dune", line 3, characters 1-12:
-  3 |  (variant x))
-       ^^^^^^^^^^^
-  Error: 'variant' is only available since version 0.1 of the experimental
-  library variants feature. Please update your dune-project file to have (using
-  library_variants 0.1).
-  [1]