+++ test/blackbox-tests/test-cases/variants/resolution-priority/bar.ml
@@ -1,3 +1,0 @@
-Direct.run ();;
-Variant.run ();;
-Test.run ();;