+++ test/blackbox-tests/test-cases/variants/variant-with-concrete-impl/lib2.default/vlib.ml
@@ -1,1 +1,0 @@
-let run () = Printf.printf "hello from lib2.default"