+++ test/blackbox-tests/test-cases/variants/variants-multi-project/prj2/vlibfoo.ml
@@ -1,1 +1,0 @@
-let implme () = print_endline "foobar"